separate mail repository
