+++ src/modules/exchange/exchange.service.ts
@@ -3,8 +3,6 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import axios from 'axios';
 import { HttpException, HttpStatus } from '@nestjs/common';
-import { MailerService } from '@nestjs-modules/mailer';
-import { Cron } from '@nestjs/schedule';
 
 import { Subscriber } from '../../db/entities/subscriber.entity';
 import { RateResponce } from './dto/exchange.dto';
@@ -25,7 +23,6 @@ export class ExchangeService {
   constructor(
     @InjectRepository(Subscriber)
     private subscriberRepository: Repository<Subscriber>,
-    private readonly mailService: MailerService,
   ) {}
 
   /**
@@ -72,30 +69,5 @@ export class ExchangeService {
     subscriber.dateCreated = new Date();
 
     return this.subscriberRepository.save(subscriber);
-  }
-
-  @Cron('00 00 12 * * *')
-  /**
-   * This function sends emails to all subscribers every day at 12 am seconds.
-   * It also can be called manually by sending a GET request to /exchange/send-emails
-   * */
-  public async sendEmails() {
-    try {
-      const { conversion_rates } = await this.getUsdUahRate();
-      const usdUahRate = conversion_rates.USD;
-      const allSubscribers = await this.subscriberRepository.find();
-      allSubscribers.forEach(subscriber => {
-        this.mailService.sendMail({
-          to: subscriber.email,
-          from: 'andrii',
-          subject: 'Current USD to UAH rate',
-          text: `Hello my dear subscriber!\nYou can buy 1 UAH for ${usdUahRate} USD, or 1 USD for ${1 / usdUahRate} UAH.\nHave a nice day!`,
-        });
-      });
-
-      return 'Emails sent';
-    } catch (error) {
-      throw new HttpException(error, HttpStatus.INTERNAL_SERVER_ERROR);
-    }
   }
 }
